add controllers, dto, exceptions, stringutils.

+++ src/main/java/com/nisum/nisumexam/repository/RolRepository.java
@@ -4,6 +4,6 @@ package com.nisum.nisumexam.repository;
 import com.nisum.nisumexam.persistence.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RolRepository extends JpaRepository<Role, Long> {
+public interface RolRepository extends JpaRepository<Role, Integer> {
 
 }
